refactor request helpers
